# api/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Face Data API",
      "version": "1.0.0",
      "description": "API for managing face data and screenshots."
    },
    "servers": [
      {
        "url": "http://localhost:3000/api",
        "description": "Local server"
      }
    ],
    "paths": {
      "/face-data": {
        "post": {
          "summary": "Create face data",
          "description": "Save face data and screenshot.",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "faceData": {
                      "type": "object",
                      "description": "Face data object"
                    },
                    "screenshot": {
                      "type": "string",
                      "description": "Base64 encoded screenshot"
                    }
                  },
                  "required": ["faceData", "screenshot"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Data saved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid data"
            },
            "500": {
              "description": "Failed to save face data or screenshot"
            }
          }
        },
        "get": {
          "summary": "Get all face data",
          "description": "Retrieve all face data files.",
          "responses": {
            "200": {
              "description": "A list of face data files",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Failed to read directory"
            }
          }
        }
      },
      "/face-data/{timestamp}": {
        "get": {
          "summary": "Get face data by timestamp",
          "description": "Retrieve specific face data file by timestamp.",
          "parameters": [
            {
              "name": "timestamp",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Face data object",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object"
                  }
                }
              }
            },
            "404": {
              "description": "File not found"
            },
            "500": {
              "description": "Failed to read file"
            }
          }
        },
        "put": {
          "summary": "Update face data by timestamp",
          "description": "Update specific face data file by timestamp.",
          "parameters": [
            {
              "name": "timestamp",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "faceData": {
                      "type": "object",
                      "description": "Face data object"
                    },
                    "screenshot": {
                      "type": "string",
                      "description": "Base64 encoded screenshot"
                    }
                  },
                  "required": ["faceData", "screenshot"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Data updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid data"
            },
            "500": {
              "description": "Failed to update face data or screenshot"
            }
          }
        },
        "delete": {
          "summary": "Delete face data by timestamp",
          "description": "Delete specific face data file by timestamp.",
          "parameters": [
            {
              "name": "timestamp",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Data deleted successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Failed to delete face data or screenshot"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "FaceData": {
          "type": "object",
          "properties": {
            "faceData": {
              "type": "object"
            },
            "screenshot": {
              "type": "string"
            }
          },
          "required": ["faceData", "screenshot"]
        }
      }
    }
  }
  